# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Forni API",
    "description": "API for the Forni bakery surplus bag platform",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:5000/api",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "paths": {
...
    "/bakery": {
      "get": {
        "tags": ["Bakeries"],
        "summary": "List all bakeries",
        "responses": {
          "200": {"description": "List of bakeries"}
        }
      },
      "post": {
        "tags": ["Bakeries"],
        "summary": "Create bakery (owner/admin)",
        "security": [{"bearerAuth": []}],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {"type": "string"},
                  "address": {"type": "string"},
                  "description": {"type": "string"}
                }
              }
            }
          }
        },
        "responses": {
          "201": {"description": "Bakery created"},
          "401": {"description": "Unauthorized"}
        }
      }
    },
    "/bakery/{bakery_id}": {
      "get": {
        "tags": ["Bakeries"],
        "summary": "Get bakery by ID",
        "parameters": [
          {"name": "bakery_id", "in": "path", "required": true, "schema": {"type": "integer"}}
        ],
        "responses": {
          "200": {"description": "Bakery details"}
        }
      },
      "put": {
        "tags": ["Bakeries"],
        "summary": "Update bakery (owner/admin)",
        "security": [{"bearerAuth": []}],
        "parameters": [
          {"name": "bakery_id", "in": "path", "required": true, "schema": {"type": "integer"}}
        ],
        "responses": {
          "200": {"description": "Bakery updated"}
        }
      },
      "delete": {
        "tags": ["Bakeries"],
        "summary": "Delete bakery (admin)",
        "security": [{"bearerAuth": []}],
        "parameters": [
          {"name": "bakery_id", "in": "path", "required": true, "schema": {"type": "integer"}}
        ],
        "responses": {
          "200": {"description": "Bakery deleted"}
        }
      }
    },
...
  }
}